[Bug Fix] In Spin + Sz-conserved

+++ src/include/mltplyMPI.h
@@ -22,20 +22,20 @@
 #include "struct.h"
 
 
-void GC_child_general_hopp_MPIdouble(int itrans, struct BindStruct *X,
+void GC_child_general_hopp_MPIdouble(unsigned long int itrans, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
 
-void GC_child_general_hopp_MPIsingle(int itrans, struct BindStruct *X,
+void GC_child_general_hopp_MPIsingle(unsigned long int itrans, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
 
-void child_general_hopp_MPIdouble(int itrans, struct BindStruct *X,
+void child_general_hopp_MPIdouble(unsigned long int itrans, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
 
-void child_general_hopp_MPIsingle(int itrans, struct BindStruct *X,
+void child_general_hopp_MPIsingle(unsigned long int itrans, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
 
-void child_general_int_spin_MPIdouble(int i_int, struct BindStruct *X,
+void child_general_int_spin_MPIdouble(unsigned long int i_int, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
 
-void child_general_int_spin_MPIsingle(int i_int, struct BindStruct *X,
+void child_general_int_spin_MPIsingle(unsigned long int i_int, struct BindStruct *X,
   double complex *tmp_v0, double complex *tmp_v1);
